[RP009] crc32_update_uint64() can be inlined
crc32_update_uint64() is defined with '__attribute__((target("...")))'.
But, if the caller and the inline function have different target attributes,
the inline function is never inlined.

Better to avoid not to block the inlining, as possible as we can.
This change uses InnoDB compile option "-mcrc32" or "-mcrc" and not use the
'__attribute__((target("...")))' for the crc32_update_uint64(),
if compiler and its version allow.